Sort 2 arrays and create their union sorted array
Implement function printarray that takes a char array and its size and outputs the
elements of the array.
Implement function swap that takes an char array and integers i and j and swaps the
ith and the jth elements of the array.

Implement function bubble sort that takes a char array and its size and sorts the input
array using the bubble sort algorithm. Modify it to print the the array at the end of
each outer loop.
Implement function insertion sort that takes a char array and its size and sorts the
input array using the insertion sort algorithm. Modify it to print the the array at the
end of each outer loop.
Implement a function union() that takes three char arrays arr1, arr2 (including their
sizes) and arr3. Two of the char arrays arr1, arr2 are already sorted in non-decreasing
order.
The function should populate the array arr3 with the union of the two sets represented
by the two arrays arr1, arr2. This array arr3 should be sorted.
E.g. union() may take the arrays arr1 = {
0a
0
,
0
c
0
,
0
c
0
,
0
f
0
,
0 k
0
,
0 m0
,
0 m0
,
0 m0
,
0
q
0
,
0
r
0} and

arr2 = {
0
c
0
,
0
f
0
,
0
f
0
,
0 h
0
,
0
r
0
,
0 x
0}, along with their sizes 10 and 6 respectively. It returns 9

and at the end of the function, arr3 = {
0a
0
,
0
c
0
,
0
f
0
,
0 h
0
,
0 k
0
,
0 m0
,
0
q
0
,
0
r
0
,
0 x
0}.

In the main() function, ask the user for the array size n1 (which should be less than
MAX SIZE) and take n1 small letters (from the alphabet) from the user. Sort them using
bubble sort. Ask the user for the array size n2 (which should be less than MAX SIZE)
and take n2 small letters (from the alphabet) from the user. Sort them using insertion
sort. Call the union() function with these two arrays as input.